requirements.txt
# -----------------------------
# Core Libraries
# -----------------------------
numpy>=1.26.5
pandas>=2.1.1
scipy>=1.11.3

# -----------------------------
# Image Processing
# -----------------------------
Pillow>=10.1.0
opencv-python>=4.9.0.76

# -----------------------------
# Deep Learning / ML
# -----------------------------
tensorflow>=2.14.0
keras>=2.14.0
scikit-learn>=1.3.1
mlflow>=2.7.0

# -----------------------------
# API Framework
# -----------------------------
fastapi>=0.113.0
uvicorn>=0.25.0
pydantic>=2.7.0
python-multipart>=0.0.6  # Required for FastAPI file uploads

# -----------------------------
# Frontend / Deployment
# -----------------------------
streamlit>=1.28.0  # Optional

# -----------------------------
# Visualization
# -----------------------------
matplotlib>=3.8.1
seaborn>=0.12.3

# -----------------------------
# Configuration / Utilities
# -----------------------------
PyYAML>=6.0
tqdm>=4.66.1



jupyter
ipykernel

plotly

pillow 